Fixed broken From behavior
Fixed problems where some smtp servers reject the From header if it's
not from the proper domain. In this case we rely on the sendmail
configuration by default. This also means the machine name info is now
being placed in the subject.

This work was tricky because of newline trimming in bash and the need to
separate the headers with newlines BUT to not have extra newlines in the
event we're not constructing a From in this script.

Added a -d|--debug switch to pass through to sendmail, for
troubleshooting.