Changes needed for the nvidia-555.58 and later driver
This version of the video driver introduced a bug where the
nvidia-auto-select mode picks the native res and as a result of this we
get a blank (black) screen. Changing the refresh to something other than
60Hz oddly fixes the issue.

The rest of this work was in figuring out how to construct the
unintuitive nvidia-settings command-line argument syntax to achieve this
without a saved conf file.