Modified/fixed HTML output
Now handling the title/pagetitle properly. Was able to remove the odd
title inside the page body by using an empty `--metadata title` and the
actual HTML title element (that shows up in the tab in a browser) to
`--variable pagetitle`

Also now assigning the filename (no path, no extension) and extension
only into variables to clarify this code.