Switched "default" margins to 1cm
This is a compromise between 2cm (which is big) and 5mm (which is very
small but I've used on some documents). A near-future change will
parameterize this with a default of 1cm.

Also added some docs clarification on using PDF to ultimately print with
lpr and changed the version to 1.2