Added arg parsing and -m/--margin switch
Added warn and die functions for error and other output handling.

Added args parsing code so we can introduce some switches. Part of this was
changing the usage from a function to a multiline string to be used by the warn
and die functions.

Added the -m/--margin switch for PDF output.